requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
alembic
python-multipart
